refactor: Refactor component and icon file structure
Changes have been made to the structure of the project files, specifically those related to UI components and icons. Component/header and icons/index files were moved up a directory level for better accessibility, and import paths were updated accordingly. This aims to reduce complexity and improve file navigation.
