+++ src/components/interface/header.tsx
@@ -3,8 +3,8 @@ import { Link, useLocation } from "@builder.io/qwik-city";
 
 import { SITENAME } from "~/constants/info";
 
+import { Navigation } from "./header/navigation";
 import styles from "./header.module.css";
-import { Navigation } from "./navigation";
 
 export const Header = component$(() => {
 	const location = useLocation();
@@ -13,7 +13,7 @@ export const Header = component$(() => {
 	return (
 		<header class={styles.header}>
 			<h1 class={styles.title}>
-				<Link aria-current={isRoot && "page"} href="/">
+				<Link aria-current={isRoot && "page"} href="/public">
 					{SITENAME}
 				</Link>
 			</h1>
