+++ src/components/interface/icons/folder-icon.tsx
@@ -1,7 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 
 import styles from "./folder-icon.module.css";
-import type { Props as IconProps } from "./index";
+import type { Props as IconProps } from "../icons";
 
 type Props = IconProps & {
 	variant?: 1 | 2 | 3;
