--- src/components/interface/icons/index.tsx
@@ -1,16 +1,0 @@
-import type { ClassList } from "@builder.io/qwik";
-
-export { FacebookIcon } from "./facebook-icon";
-export { FolderIcon } from "./folder-icon";
-export { GamepadIcon } from "./gamepad-icon";
-export { GitHubIcon } from "./github-icon";
-export { MailIcon } from "./mail-icon";
-export { MarkdownIcon } from "./markdown-icon";
-export { PhoneIcon } from "./phone-icon";
-export { TerminalIcon } from "./terminal-icon";
-export { UserIcon } from "./user-icon";
-export { XIcon } from "./x-icon";
-
-export type Props = {
-	class?: ClassList;
-};
