+++ src/routes/layout.tsx
@@ -1,7 +1,7 @@
 import { component$, Slot } from "@builder.io/qwik";
 
 import { Footer } from "~/components/interface/footer";
-import { Header } from "~/components/interface/header/header";
+import { Header } from "~/components/interface/header";
 import { useGlobalStyles } from "~/hooks/global-styles/global-styles";
 
 /** @private */
